\name{pBrackets-package}
\alias{pBrackets-package}
\alias{pBrackets}
\docType{package}
\title{
Plot Brackets
}
\description{
Adds different kinds of brackets to a plot, including braces, chevrons,
parentheses or square brackets.
}
\details{
\tabular{ll}{
Package: \tab pBrackets\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2021-05-17\cr
License: \tab GPL version 2 or newer \cr
}
}
\author{
Andreas Schulz
Maintainer: Andreas Schulz <ades-s@web.de>
}
\keyword{braces}
\keyword{brackets}
